feat : main 페이지 Skeleton UI
