+++ src/components/Main/MainLeft.jsx
@@ -1,42 +1,11 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import styled from "styled-components";
 
 import commonBanner from "../../assets/common-banner.svg";
-import { useRecoilValue } from "recoil";
-import { accountname } from "../../recoil";
-import { followingAPI, productListAPI } from "../../api";
 import { checkImageUrl } from "../../utils";
 import { Link } from "react-router-dom";
 
-export default function MainLeft() {
-  const accountName = useRecoilValue(accountname);
-  const [render, setRender] = useState([]);
-
-  useEffect(() => {
-    const fetchfollowProduct = async () => {
-      try {
-        const usernameResponse = await followingAPI(accountName);
-        const renderProduct = usernameResponse.data.slice(0, 4);
-
-        const imgPromises = renderProduct.map(async (item) => {
-          const response = await productListAPI({ accountname: item.accountname });
-
-          if (response.product) {
-            const img = response.product[0].itemImage.split(",")[0];
-            const id = response.product[0].id;
-            return { img, id };
-          }
-        });
-
-        const imgIdResults = await Promise.all(imgPromises);
-        setRender(imgIdResults);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    fetchfollowProduct();
-  }, []);
-
+export default function MainLeft({ render }) {
   return (
     <MainLeftSection>
       <h2>
@@ -44,11 +13,14 @@ export default function MainLeft() {
       </h2>
 
       <article>
-        {render.map((item, index) => (
-          <Link key={item.id} to={`/productDetail/${item.id}`}>
-            <img src={checkImageUrl(item.img, "post")} alt={`팔로우한 작가 상품 이미지 ${index}`} />
-          </Link>
-        ))}
+        {render
+          .filter((i) => i !== undefined)
+          .map((item, index) => (
+            <Link key={item.id} to={`/productDetail/${item.id}`}>
+              <img src={checkImageUrl(item.img, "post")} alt={`팔로우한 작가 상품 이미지 ${index}`} />
+            </Link>
+          ))
+          .slice(0, 4)}
       </article>
     </MainLeftSection>
   );
